requirements.txt
# Core ML Frameworks
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Optional: For development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0